Split out Illuminate and Respite sources to their own repositories
