+++ sources/illuminate/src/Resolvers/HasOne.php
