+++ sources/respite/composer.json
@@ -1,0 +1,36 @@
+{
+  "name": "sprocketbox/articulate-respite",
+  "description": "An Articulate source providing Respite support",
+  "type": "library",
+  "license": "MIT",
+  "version": "dev-develop",
+  "authors": [
+    {
+      "name": "Ollie Read",
+      "email": "ollie@sprocketbox.io"
+    }
+  ],
+  "require": {
+    "php": "^7.2",
+    "sprocketbox/articulate": "1",
+    "sprocketbox/respite": "1",
+  },
+  "provide": {
+    "sprocketbox/articulate-source": "1.0.0"
+  },
+  "autoload": {
+    "psr-4": {
+      "Sprocketbox\\Articulate\\Sources\\Respite": "src/"
+    }
+  },
+  "extra": {
+    "laravel": {
+      "providers": [
+        "Sprocketbox\\Articulate\\Sources\\Respite\\ServiceProvider"
+      ]
+    }
+  },
+  "suggest": {
+    "sprocketbox/respite": "Allows for the usage of RESTful APIs as a data source"
+  }
+}
