+++ sources/respite/src/ServiceProvider.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Sprocketbox\Articulate\Sources\Respite;
+
+use Illuminate\Support\ServiceProvider as BaseServiceProvider;
+
+class ServiceProvider extends BaseServiceProvider
+{
+    public function register()
+    {
+        entities()->registerSource(new RespiteSource);
+    }
+}
